Implimented ability Artillery Blast. Assigned to the generator button for now in the abillities menu. Seems to work fine, I am aware there is an error in the console when it is used, but it is not breaking anything right now and I will fix it later.

+++ Assets/Scripts/Abilities/Ability.cs
@@ -10,7 +10,7 @@ public abstract class Ability : ScriptableObject, ICollectible
     public AudioClip sound;
     public Sprite sprite;
     public float baseCoolDown = 1f;
-    public float targetRadius = 10f;
+    public int targetRadius = 10;
     public int powerCost;
     public int duration;
